refactor(datagrid-web): move code to plugin
