+++ packages/shared/widget-plugin-grid/src/event-switch/use-event-switch.ts
