+++ packages/shared/widget-plugin-grid/src/selection/types.ts
@@ -1,4 +1,4 @@
-import { SelectionMultiValue, SelectionSingleValue } from "mendix";
+import { SelectionMultiValue, SelectionSingleValue, ObjectItem } from "mendix";
 
 export type SelectionType = "None" | "Single" | "Multi";
 
@@ -7,3 +7,5 @@ export type MultiSelectionStatus = "none" | "all" | "some";
 export type SelectionStatus = "none" | "all" | "some" | "unknown";
 
 export type WidgetSelectionProperty = SelectionSingleValue | SelectionMultiValue | string | undefined;
+
+export type SelectFx = (item: ObjectItem, shiftKey: boolean) => void;
